refactor: 중복되는 dto삭제 및 통합
